+++ plugins/catalog-backend-module-github/src/lib/org.ts
@@ -14,7 +14,11 @@
  * limitations under the License.
  */
 
-import { GroupEntity, UserEntity } from '@backstage/catalog-model';
+import {
+  DEFAULT_NAMESPACE,
+  GroupEntity,
+  UserEntity,
+} from '@backstage/catalog-model';
 
 export function buildOrgHierarchy(groups: GroupEntity[]) {
   const groupsByName = new Map(groups.map(g => [g.metadata.name, g]));
@@ -52,8 +56,19 @@ export function buildOrgHierarchy(groups: GroupEntity[]) {
 // Ensure that users have their direct group memberships.
 export function assignGroupsToUsers(
   users: UserEntity[],
-  groupMemberUsers: Map<string, string[]>,
+  groups: GroupEntity[],
 ) {
+  const groupMemberUsers = new Map(
+    groups.map(group => {
+      const groupKey =
+        group.metadata.namespace &&
+        group.metadata.namespace !== DEFAULT_NAMESPACE
+          ? `${group.metadata.namespace}/${group.metadata.name}`
+          : group.metadata.name;
+      return [groupKey, group.spec.members || []];
+    }),
+  );
+
   const usersByName = new Map(users.map(u => [u.metadata.name, u]));
   for (const [groupName, userNames] of groupMemberUsers.entries()) {
     for (const userName of userNames) {