+++ plugins/explore/src/components/DefaultExplorePage/DefaultExplorePage.test.tsx
@@ -21,8 +21,8 @@ import React from 'react';
 import { DefaultExplorePage } from './DefaultExplorePage';
 
 describe('<DefaultExplorePage />', () => {
-  const catalogApi: jest.Mocked<typeof catalogApiRef.T> = {
-    addLocation: jest.fn(_a => new Promise(() => {})),
+  const catalogApi = {
+    addLocation: jest.fn(),
     getEntities: jest.fn(),
     getLocationByRef: jest.fn(),
     getLocationById: jest.fn(),