+++ packages/test-utils/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @backstage/test-utils
 
+## 1.2.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-app-api@1.2.0
+  - @backstage/core-plugin-api@1.1.0
+  - @backstage/plugin-permission-common@0.7.1
+  - @backstage/types@1.0.1
+  - @backstage/config@1.0.4
+  - @backstage/theme@0.2.16
+  - @backstage/plugin-permission-react@0.4.7
+
 ## 1.2.2-next.0
 
 ### Patch Changes