+++ plugins/auth-node/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @backstage/plugin-auth-node
 
+## 0.2.7
+
+### Patch Changes
+
+- 7573b65232: Internal refactor of imports to avoid circular dependencies
+- Updated dependencies
+  - @backstage/backend-common@0.16.0
+  - @backstage/config@1.0.4
+  - @backstage/errors@1.1.3
+
+## 0.2.7-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.16.0-next.1
+  - @backstage/config@1.0.4-next.0
+  - @backstage/errors@1.1.3-next.0
+
 ## 0.2.7-next.0
 
 ### Patch Changes