+++ plugins/tech-insights-backend-module-jsonfc/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @backstage/plugin-tech-insights-backend-module-jsonfc
 
+## 0.1.22
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.16.0
+  - @backstage/plugin-tech-insights-node@0.3.6
+  - @backstage/config@1.0.4
+  - @backstage/errors@1.1.3
+  - @backstage/plugin-tech-insights-common@0.2.8
+
+## 0.1.22-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.16.0-next.1
+  - @backstage/plugin-tech-insights-node@0.3.6-next.1
+  - @backstage/config@1.0.4-next.0
+  - @backstage/errors@1.1.3-next.0
+  - @backstage/plugin-tech-insights-common@0.2.8-next.0
+
 ## 0.1.22-next.0
 
 ### Patch Changes