+++ plugins/techdocs-addons-test-utils/CHANGELOG.md
@@ -1,5 +1,44 @@
 # @backstage/plugin-techdocs-addons-test-utils
 
+## 1.0.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-techdocs@1.4.0
+  - @backstage/core-components@0.12.0
+  - @backstage/core-app-api@1.2.0
+  - @backstage/core-plugin-api@1.1.0
+  - @backstage/plugin-techdocs-react@1.0.6
+  - @backstage/plugin-catalog@1.6.1
+  - @backstage/integration-react@1.1.6
+  - @backstage/plugin-search-react@1.2.1
+  - @backstage/test-utils@1.2.2
+  - @backstage/theme@0.2.16
+
+## 1.0.6-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-techdocs@1.4.0-next.2
+  - @backstage/core-components@0.12.0-next.1
+  - @backstage/core-app-api@1.2.0-next.0
+  - @backstage/core-plugin-api@1.1.0-next.0
+  - @backstage/integration-react@1.1.6-next.1
+  - @backstage/test-utils@1.2.2-next.0
+  - @backstage/theme@0.2.16
+  - @backstage/plugin-catalog@1.6.1-next.1
+  - @backstage/plugin-search-react@1.2.1-next.1
+  - @backstage/plugin-techdocs-react@1.0.6-next.1
+
+## 1.0.6-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-techdocs@1.4.0-next.1
+
 ## 1.0.6-next.0
 
 ### Patch Changes